Always run the terraform generator on new apps. No longer optional
