+++ lib/generators/rails_template18f/circleci/templates/circleci/config.yml.tt
@@ -3,8 +3,8 @@ version: 2.1
 orbs:
   ruby: circleci/ruby@2.1.3
   node: circleci/node@5.2.0
-  browser-tools: circleci/browser-tools@1.4.8<% if terraform? %>
-  terraform: circleci/terraform@3.2.1<% end %>
+  browser-tools: circleci/browser-tools@1.4.8
+  terraform: circleci/terraform@3.2.1
 
 commands:
   setup-project:
@@ -30,53 +30,6 @@ commands:
       - run:
           name: Run trestle command
           command: docker run -u "$(id -u):$(id -g)" -v $(pwd)/doc/compliance/oscal:/app/docs:rw ghcr.io/gsa-tts/trestle:<< parameters.tag >> << parameters.cmd >><% end %>
-  cg-deploy:
-    description: "Login to cloud foundry space with service account credentials
-      and push application using deployment configuration file."
-    parameters:
-      cloudgov_username:
-        description: "Name of CircleCI project environment variable that
-          holdes deployer username for cloudgov space"
-        type: env_var_name
-      cloudgov_password:
-        description: "Name of CircleCI project environment variable that
-          holds deployer password for cloudgov space"
-        type: env_var_name
-      cloudgov_org:
-        description: "cloud.gov organization name"
-        type: string
-      cloudgov_space:
-        description: "cloud.gov space name"
-        type: string
-      deploy_config_file:
-        description: "Path to deployment configuration file"
-        type: string
-      rails_master_key:
-        description: "Name of CircleCI project environment variable holding the RAILS_MASTER_KEY"
-        type: env_var_name
-    steps:
-      - run:
-          name: Vendor gems
-          command: bundle cache --all
-      - run:
-          name: Install Cloud Foundry CLI
-          command: |
-            curl -v -L -o cf-cli_amd64.deb 'https://packages.cloudfoundry.org/stable?release=debian64&version=v8&source=github'
-            sudo dpkg -i cf-cli_amd64.deb
-      - run:
-          name: Login with service account
-          command: |
-            cf login -a api.fr.cloud.gov \
-              -u ${<< parameters.cloudgov_username >>} \
-              -p ${<< parameters.cloudgov_password >>} \
-              -o << parameters.cloudgov_org >> \
-              -s << parameters.cloudgov_space >>
-      - run:
-          name: Push application with deployment vars
-          command: |
-            cf push --strategy rolling \
-              --vars-file << parameters.deploy_config_file >> \
-              --var rails_master_key=${<< parameters.rails_master_key >>}
 
 jobs:
   build:
@@ -300,7 +253,6 @@ jobs:
       - run:
           name: Run pa11y-ci
           command: yarn run pa11y-ci -c pa11yci.js
-<% if terraform? %>
   terraform_plan_staging:
     executor: terraform/default
     steps:
@@ -357,31 +309,6 @@ jobs:
           at: .
       - terraform/apply:
           path: terraform/production
-<% end %>
-  deploy_staging:
-    docker:
-      - image: cimg/ruby:<%= ruby_version %>
-    steps:
-      - setup-project
-      - cg-deploy:
-          cloudgov_username: CF_STAGING_USERNAME
-          cloudgov_password: CF_STAGING_PASSWORD
-          cloudgov_org: <%= cloud_gov_organization %>
-          cloudgov_space: <%= cloud_gov_staging_space %>
-          deploy_config_file: config/deployment/staging.yml
-          rails_master_key: RAILS_MASTER_KEY
-  deploy_production:
-    docker:
-      - image: cimg/ruby:<%= ruby_version %>
-    steps:
-      - setup-project
-      - cg-deploy:
-          cloudgov_username: CF_PRODUCTION_USERNAME
-          cloudgov_password: CF_PRODUCTION_PASSWORD
-          cloudgov_org: <%= cloud_gov_organization %>
-          cloudgov_space: <%= cloud_gov_production_space %>
-          deploy_config_file: config/deployment/production.yml
-          rails_master_key: PRODUCTION_RAILS_MASTER_KEY
 
 workflows:
   version: 2.1
@@ -409,7 +336,7 @@ workflows:
             - build
       - a11y_scan:
           requires:
-            - build<% if terraform? %>
+            - build
       - terraform_plan_staging:
           filters:
             branches:
@@ -433,27 +360,7 @@ workflows:
             branches:
               only: production
           requires:
-            - approve_production_terraform<% end %>
-      - deploy_staging:
-          filters:
-            branches:
-              only: main
-          requires:
-            - test
-            - static_security_scans
-            - owasp_scan
-            - a11y_scan<% if terraform? %>
-            - terraform_apply_staging<% end %>
-      - deploy_production:
-          filters:
-            branches:
-              only: production
-          requires:
-            - test
-            - static_security_scans
-            - owasp_scan
-            - a11y_scan<% if terraform? %>
-            - terraform_apply_production<% end %>
+            - approve_production_terraform
   daily_scan:
     triggers:
       - schedule:
