+++ lib/generators/rails_template18f/github_actions/templates/github/dependabot.yml.tt
@@ -14,7 +14,7 @@ updates:
   directory: "/"
   schedule:
     interval: daily
-  open-pull-requests-limit: 10<% if terraform? %>
+  open-pull-requests-limit: 10
 - package-ecosystem: terraform
   directories:
     - "/terraform/production"
@@ -22,4 +22,3 @@ updates:
   schedule:
     interval: weekly
   open-pull-requests-limit: 10
-<% end %>
