+++ lib/generators/rails_template18f/terraform/templates/terraform/production.tfvars.tt
@@ -4,3 +4,6 @@ rds_plan_name      = "TKTK-production-rds-plan"
 <% if has_active_job? %>redis_plan_name    = "TKTK-production-redis-plan"<% end %>
 custom_domain_name = null
 optional_host_name = null
+web_instances      = 2
+web_memory         = "512M"
+<% if has_active_job? %>worker_memory      = "512M"<% end %>
