+++ lib/rails_template18f/generators.rb
@@ -7,6 +7,5 @@ module RailsTemplate18f
     autoload :Base
     autoload :CloudGovOptions
     autoload :CloudGovParsing
-    autoload :PipelineOptions
   end
 end
