--- templates/config/deployment/production.yml
@@ -1,3 +1,0 @@
-env: production
-web_instances: 2
-web_memory: 512M
