--- templates/config/deployment/staging.yml
@@ -1,3 +1,0 @@
-env: staging
-web_instances: 1
-web_memory: 256M
