--- templates/manifest.yml.tt
@@ -1,17 +1,0 @@
----
-applications:
-- name: <%= app_name %>-((env))
-  buildpacks:
-    - ruby_buildpack
-  env:
-    RAILS_MASTER_KEY: ((rails_master_key))
-    RAILS_ENV: ((env))
-    RAILS_LOG_TO_STDOUT: true
-    RAILS_SERVE_STATIC_FILES: true
-  processes:
-  - type: web
-    instances: ((web_instances))
-    memory: ((web_memory))
-    command: bundle exec rake cf:on_first_instance db:migrate && exec bundle exec rails s -b 0.0.0.0 -p $PORT -e $RAILS_ENV
-  services:
-  - <%= app_name %>-rds-((env))
